+++ source/extensions/omni.isaac.core/python/tests/test_particle_system.py
@@ -10,7 +10,7 @@
 import omni.kit.test
 from omni.isaac.core import World
 from omni.isaac.core.prims.soft.particle_system import ParticleSystem
-from omni.isaac.core.tests.test_semantics import TestProperties
+from omni.isaac.core.tests.common import TestProperties
 
 # NOTE:
 #   omni.kit.test - std python's unittest module with additional wrapping to add support for async/await tests
@@ -56,7 +56,6 @@ class TestParticleSystem(omni.kit.test.AsyncTestCase, TestProperties):
 
         self.my_world.scene.add(self.particle_system)
         await self.my_world.reset_async(soft=False)
-        await update_stage_async()
         await self.my_world.stop_async()
 
         await self.scalar_prop_test(